Move default templates to RTKQ API + constant
